fix: error in WriteHTML() ('The system cannot find the path specified')
If the output directory had been created previously, the program would
catch an error indicating that the directory already existed
(`os.IsExists()` returns `true` or equivalent value) and call
`os.RemoveAll()` to remove the directory and all its children. In this
case, the directory would need to be re-created, otherwise errors
related to nonexistent paths would be raised. Now, the program performs
a deep removal of the target directory prior to attempting to create the
output directory.