fix: template image would not be replaced when center-aligned
Template image would not be replaced by the target image in the output
Word Document when center-aligning the images; however, the template
image would be replaced by the target image in the output Word Document
when left-aligning the images. The solution was to search for the
original image (to be replaced) by filename with both the .jpg and .jpeg
extensions. The image replacement method still functions properly since
.jpg and .jpeg are synonymous for the same file type.